requirements.txt
pypdf==6.7.5
pytest==8.3.5
numpy==2.4.2
Pillow==12.0.0
